refactor(styles-processor): rename file to style-processor
