+++ lib/broccoli-style-processor.js
@@ -5,7 +5,7 @@
 const Filter = require('broccoli-persistent-filter');
 const extend = require('deep-extend');
 const path = require('path');
-const processStyles = require('./styles-processor');
+const processStyles = require('./style-processor');
 
 class BroccoliStyleProcessor extends Filter {
 	constructor(inputNodes, options) {
