+++ lib/style-processor.js
